+++ packages/stream/src/group.js
@@ -1,30 +1,36 @@
 const through = require('./through');
 
-function group({keyFn, eager}) {
+function group({keyFn, eager, maxCount}) {
   let groups = {};
   let lastKey = false;
+  let count = 0
+
+  const flush = (readable) => {
+    Object.values(groups)
+      .filter(group => group.length)
+      .forEach(group => readable.push(group));
+    groups = {}
+    lastKey = false
+    count = 0
+  }
 
   return through(function (entry, _, cb) {
     const key = keyFn(entry);
-    if (lastKey === false) {
-      groups[key] = [entry];
-    } else if (eager && lastKey !== key) {
-      this.push(groups[lastKey]);
-      delete groups[lastKey];
-      groups[key] = [entry];
-    } else {
-      if (!groups[key]) {
-        groups[key] = [];
-      }
-      groups[key].push(entry);
+    if (eager && lastKey && lastKey !== key) {
+      flush(this);
     }
+    if (!groups[key]) {
+      groups[key] = [];
+    }
+    groups[key].push(entry);
     lastKey = key;
+    count++
+    if (maxCount > 0 && count == maxCount) {
+      flush(this);
+    }
     cb();
   }, function(cb) {
-    const that = this;
-    Object.values(groups)
-      .filter(group => group.length)
-      .forEach(group => that.push(group));
+    flush(this);
     cb();
   });